# experian-single-signer/rest/bin/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: This API allows a caller to request summarized or detailed information about a company's debts.
  version: 1.0.0
  title: Business Debts Records
  termsOfService: https://dev-us-api.experian.com/terms-of-service
  contact:
    name: API Team
    url: https://dev-us-api.experian.com/contact
    email: apiteam@br.experian.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: dev-us-api.experian.com
basePath: "/businessinformation/businessdebtsrecords/v1"
schemes:
  - https
securityDefinitions:
  OauthSecurity:
    type: oauth2
    flow: password
    tokenUrl: https://dev-us-api.experian.com/br/oauth/v1/token
    scopes: 
      user: "user Scope"
paths:
  "/cms/signatures/validate":
    post:
      summary: Validador de assinaturas.
      description: Realiza a validação da assinatura de um documento.
      operationId: validate
      tags:
        - Validador de assinaturas
      produces:
        - application/json
      responses:
        '401':
          description: Invalid or Expired Access Token
        '500':
          description: Invalid companyId 
        '200':
          description: OK
      
      parameters:
        - name: Content-Type
          in: header
          default: application/json
          description: Content-Type field is to describe the data contained in the body.
          required: true
          type: string
        - name: Authorization
          in: header
          description: The client sends the access token preceded by Bearer.
          required: true
          type: string
        - in: "body"
          name: "validationRequest"
          description: "Objeto que contém o documento que será validado"
          required: true
          schema:
            $ref: "#/definitions/Validation"

  "/debtssummaries/{companyId}":
    get:
      summary: List of company's debts in summarized form.
      description: Retrieve summarized information about debts occurrences of a given company grouped by debts types, total debts amount and quantity of debts.
      operationId: retrieveDebtsSummaries
      tags:
        - Debts Summaries
      produces:
        - application/json
      responses:
        '401':
          description: Invalid or Expired Access Token
        '500':
          description: Invalid companyId 
        '200':
          description: OK
          schema: 
            type: 'array'
            items:
              $ref: '#/definitions/Resumo'
      parameters:
        - name: Content-Type
          in: header
          default: application/json
          description: Content-Type field is to describe the data contained in the body.
          required: true
          type: string
        - name: Authorization
          in: header
          description: The client sends the access token preceded by Bearer.
          required: true
          type: string
        - name: companyId
          in: path
          description: Company ID
          required: true
          type: integer
          format: int64
  "/olaMundo/{msg}":
    get:
      summary: List of company's debts in summarized form.
      description: Retrieve summarized information about debts occurrences of a given company grouped by debts types, total debts amount and quantity of debts.
      operationId: getMsg
      tags:
        - Ola mundo
      produces:
        - application/json
      responses:
        '401':
          description: Invalid or Expired Access Token
        '500':
          description: Invalid companyId 
        '200':
          description: OK
          schema: 
            type: 'array'
            items:
              $ref: '#/definitions/OlaMundoMsg'
      parameters:
        - name: msg
          in: path
          description: Company ID
          required: true
          type: string
definitions:
  Resumo:
    description: ''
    required:
      - updateDate
      - quantity
      - total
      - type
    properties:
      updateDate:
        type: string
        format: date-time
      quantity:
        type: number
      total:
        type: number
      type:
        type: string
  OlaMundoMsg:
      description: ''
      required:
        - mensagem
      properties:
        mensagem:
          type: string
  formatoDocumento:
        type: "string"
        description: "Formatos aceitos para assinatura"
        enum:
        - "String"
        - "Hash"
        - "Hash64"
        - "P7s"
        - "Xml"
        - "Doc"
        - "Docx"
        - "Txt"
        - "Csv"
        - "Xls"
        - "Xlsx"
        - "Xlsm"
  Validation:
    description: 'Objeto utilizado para encapsular uma assinatura que será validada'
    required:
      - assinatura
      - conteudo
    properties:
      assinatura:
        type: string
      conteudo:
        type: string
  Conteudo:
    description: 'Informação que será validada'
    required:
      - attached
      - type
      - descricao
    properties:
      attached:
        type: boolean
      type:
        $ref: '#/definitions/formatoDocumento'
      descricao:
        type: string