# experian-single-signer/rest/target/generated-sources/swagger/gen/br/com/experian/single/signer/rest/model/Conteudo.java
/*
 * Business Debts Records
 * This API allows a caller to request summarized or detailed information about a company's debts.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@br.experian.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package br.com.experian.single.signer.rest.model;

import java.util.Objects;
import br.com.experian.single.signer.rest.model.FormatoDocumento;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;

/**
 * Informação que será validada
 */
@ApiModel(description = "Informação que será validada")
public class Conteudo   {
  @JsonProperty("attached")
  private Boolean attached = null;

  @JsonProperty("type")
  private FormatoDocumento type = null;

  @JsonProperty("descricao")
  private String descricao = null;

  public Conteudo attached(Boolean attached) {
    this.attached = attached;
    return this;
  }

  /**
   * Get attached
   * @return attached
   **/
  @JsonProperty("attached")
  @ApiModelProperty(required = true, value = "")
  public Boolean getAttached() {
    return attached;
  }

  public void setAttached(Boolean attached) {
    this.attached = attached;
  }

  public Conteudo type(FormatoDocumento type) {
    this.type = type;
    return this;
  }

  /**
   * Get type
   * @return type
   **/
  @JsonProperty("type")
  @ApiModelProperty(required = true, value = "")
  public FormatoDocumento getType() {
    return type;
  }

  public void setType(FormatoDocumento type) {
    this.type = type;
  }

  public Conteudo descricao(String descricao) {
    this.descricao = descricao;
    return this;
  }

  /**
   * Get descricao
   * @return descricao
   **/
  @JsonProperty("descricao")
  @ApiModelProperty(required = true, value = "")
  public String getDescricao() {
    return descricao;
  }

  public void setDescricao(String descricao) {
    this.descricao = descricao;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Conteudo conteudo = (Conteudo) o;
    return Objects.equals(this.attached, conteudo.attached) &&
        Objects.equals(this.type, conteudo.type) &&
        Objects.equals(this.descricao, conteudo.descricao);
  }

  @Override
  public int hashCode() {
    return Objects.hash(attached, type, descricao);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Conteudo {\n");
    
    sb.append("    attached: ").append(toIndentedString(attached)).append("\n");
    sb.append("    type: ").append(toIndentedString(type)).append("\n");
    sb.append("    descricao: ").append(toIndentedString(descricao)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}