# experian-single-signer/rest/target/generated-sources/swagger/gen/br/com/experian/single/signer/rest/model/OlaMundoMsg.java
/*
 * Business Debts Records
 * This API allows a caller to request summarized or detailed information about a company's debts.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@br.experian.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package br.com.experian.single.signer.rest.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;

/**
 * 
 */
@ApiModel(description = "")
public class OlaMundoMsg   {
  @JsonProperty("mensagem")
  private String mensagem = null;

  public OlaMundoMsg mensagem(String mensagem) {
    this.mensagem = mensagem;
    return this;
  }

  /**
   * Get mensagem
   * @return mensagem
   **/
  @JsonProperty("mensagem")
  @ApiModelProperty(required = true, value = "")
  public String getMensagem() {
    return mensagem;
  }

  public void setMensagem(String mensagem) {
    this.mensagem = mensagem;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    OlaMundoMsg olaMundoMsg = (OlaMundoMsg) o;
    return Objects.equals(this.mensagem, olaMundoMsg.mensagem);
  }

  @Override
  public int hashCode() {
    return Objects.hash(mensagem);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class OlaMundoMsg {\n");
    
    sb.append("    mensagem: ").append(toIndentedString(mensagem)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}