# experian-single-signer/rest/target/generated-sources/swagger/gen/br/com/experian/single/signer/rest/model/Resumo.java
/*
 * Business Debts Records
 * This API allows a caller to request summarized or detailed information about a company's debts.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@br.experian.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package br.com.experian.single.signer.rest.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.math.BigDecimal;
import java.util.Date;

/**
 * 
 */
@ApiModel(description = "")
public class Resumo   {
  @JsonProperty("updateDate")
  private Date updateDate = null;

  @JsonProperty("quantity")
  private BigDecimal quantity = null;

  @JsonProperty("total")
  private BigDecimal total = null;

  @JsonProperty("type")
  private String type = null;

  public Resumo updateDate(Date updateDate) {
    this.updateDate = updateDate;
    return this;
  }

  /**
   * Get updateDate
   * @return updateDate
   **/
  @JsonProperty("updateDate")
  @ApiModelProperty(required = true, value = "")
  public Date getUpdateDate() {
    return updateDate;
  }

  public void setUpdateDate(Date updateDate) {
    this.updateDate = updateDate;
  }

  public Resumo quantity(BigDecimal quantity) {
    this.quantity = quantity;
    return this;
  }

  /**
   * Get quantity
   * @return quantity
   **/
  @JsonProperty("quantity")
  @ApiModelProperty(required = true, value = "")
  public BigDecimal getQuantity() {
    return quantity;
  }

  public void setQuantity(BigDecimal quantity) {
    this.quantity = quantity;
  }

  public Resumo total(BigDecimal total) {
    this.total = total;
    return this;
  }

  /**
   * Get total
   * @return total
   **/
  @JsonProperty("total")
  @ApiModelProperty(required = true, value = "")
  public BigDecimal getTotal() {
    return total;
  }

  public void setTotal(BigDecimal total) {
    this.total = total;
  }

  public Resumo type(String type) {
    this.type = type;
    return this;
  }

  /**
   * Get type
   * @return type
   **/
  @JsonProperty("type")
  @ApiModelProperty(required = true, value = "")
  public String getType() {
    return type;
  }

  public void setType(String type) {
    this.type = type;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Resumo resumo = (Resumo) o;
    return Objects.equals(this.updateDate, resumo.updateDate) &&
        Objects.equals(this.quantity, resumo.quantity) &&
        Objects.equals(this.total, resumo.total) &&
        Objects.equals(this.type, resumo.type);
  }

  @Override
  public int hashCode() {
    return Objects.hash(updateDate, quantity, total, type);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Resumo {\n");
    
    sb.append("    updateDate: ").append(toIndentedString(updateDate)).append("\n");
    sb.append("    quantity: ").append(toIndentedString(quantity)).append("\n");
    sb.append("    total: ").append(toIndentedString(total)).append("\n");
    sb.append("    type: ").append(toIndentedString(type)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}