# experian-single-signer/rest/target/generated-sources/swagger/gen/br/com/experian/single/signer/rest/model/Validation.java
/*
 * Business Debts Records
 * This API allows a caller to request summarized or detailed information about a company's debts.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@br.experian.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package br.com.experian.single.signer.rest.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;

/**
 * Objeto utilizado para encapsular uma assinatura que será validada
 */
@ApiModel(description = "Objeto utilizado para encapsular uma assinatura que será validada")
public class Validation   {
  @JsonProperty("assinatura")
  private String assinatura = null;

  @JsonProperty("conteudo")
  private String conteudo = null;

  public Validation assinatura(String assinatura) {
    this.assinatura = assinatura;
    return this;
  }

  /**
   * Get assinatura
   * @return assinatura
   **/
  @JsonProperty("assinatura")
  @ApiModelProperty(required = true, value = "")
  public String getAssinatura() {
    return assinatura;
  }

  public void setAssinatura(String assinatura) {
    this.assinatura = assinatura;
  }

  public Validation conteudo(String conteudo) {
    this.conteudo = conteudo;
    return this;
  }

  /**
   * Get conteudo
   * @return conteudo
   **/
  @JsonProperty("conteudo")
  @ApiModelProperty(required = true, value = "")
  public String getConteudo() {
    return conteudo;
  }

  public void setConteudo(String conteudo) {
    this.conteudo = conteudo;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Validation validation = (Validation) o;
    return Objects.equals(this.assinatura, validation.assinatura) &&
        Objects.equals(this.conteudo, validation.conteudo);
  }

  @Override
  public int hashCode() {
    return Objects.hash(assinatura, conteudo);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Validation {\n");
    
    sb.append("    assinatura: ").append(toIndentedString(assinatura)).append("\n");
    sb.append("    conteudo: ").append(toIndentedString(conteudo)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}